PIn.Int changed to PIn.Long in PatPlanOrdinalTwoToOne. Backported to beta and stable. Was causing "Value was either too large or too small for Int32" error.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@7540 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Misc/DatabaseMaintenance.cs
@@ -1883,7 +1883,7 @@ namespace OpenDentBusiness {
 			else {
 				int numberFixed=0;
 				for(int i=0;i<table.Rows.Count;i++) {
-					PatPlan patPlan=PatPlans.GetPatPlan(PIn.Int(table.Rows[i][1].ToString()),0);
+					PatPlan patPlan=PatPlans.GetPatPlan(PIn.Long(table.Rows[i][1].ToString()),0);
 					if(patPlan!=null) {//Unlikely but possible if plan gets deleted by a user during this check.
 						PatPlans.SetOrdinal(patPlan.PatPlanNum,1);
 						numberFixed++;
